Update readme for clarity and added git info
We want the instructions to be as clear as possible so that candidates don't unnecessarily get hung up on certain details and focus on the right aspects of the challenge.

Additionally, we want candidates to show their git skills if they have any, so we're adding instructions to create a git repo for the submission if the candidate knows and can use git to do so.