Make the README more concise
The instructions could be clearer, in my opinion, so I'm hoping that
making the README more concise does precisely that.

Additionally, I feel as though some candidates don't read through the
entire document based on some of their questions and feedback, and
that's really not their fault as the README is exceptionally long for a
"simple" take home challenge. Shortening the README should hopefully
ensure that candidates are reading through the entirety of the document
before they get started.

Plus, who wants to read that much text for a take home challenge when
candidates are likely still working a full-time job and possibly
interviewing with other companies concurrently? Let's keep things short
and simple for our candidates.